implemented Router and Switch. created components folder. renamed media' folder to 'assets'.
