+++ plants/src/App.js
@@ -1,46 +1,55 @@
 import './App.css';
-import React from 'react'
-import Plant from './Plant';
-import Header from './Header'
-import ReactDOM from 'react-dom';
-import { CssBaseline, AppBar, Toolbar, Typography, Button, Box } from '@mui/material'
-import GrassIcon from '@mui/icons-material/Grass'
-import MainCard from './MainCard';
-import LoginForm from './loginForm';
-import RegisterForm from './registerForm';
+import React from 'react';
+import { Route, Switch, Redirect } from 'react-router-dom';
+import Plant from './components/Plant';
+import Header from './components/Header';
+import { CssBaseline, AppBar, Toolbar, Typography, Button, Box } from '@mui/material';
+import MainCard from './components/MainCard';
+import LoginForm from './components/loginForm';
+import RegisterForm from './components/registerForm';
 import axios from "axios";
+
 const appStyle = {
-	height: '250px',
-  	display: 'flex'
+    height: '250px',
+    display: 'flex'
 };
 
-function App() {
+function App()
+{
 
-    async function handleRegisterSubmit(data) {
-      const res = await axios.post('https://build-week-water-my-plants-1.herokuapp.com/api/auth/register', data);
-      console.log(res);
+    async function handleRegisterSubmit(data)
+    {
+        const res = await axios.post('https://build-week-water-my-plants-1.herokuapp.com/api/auth/register', data);
+        console.log(res);
     }
-    async function handleLoginSubmit(data) {
-      const res = await axios.post('https://build-week-water-my-plants-1.herokuapp.com/api/auth/login', data);
-      console.log(res);
+    async function handleLoginSubmit(data)
+    {
+        const res = await axios.post('https://build-week-water-my-plants-1.herokuapp.com/api/auth/login', data);
+        console.log(res);
     }
-  return (
-    <>
-      <CssBaseline />
+    return (
+        <>
+            <CssBaseline />
 
-        <Header />
-        <main>
-          <Box sx={{display: "flex", flexDirection: "column", alignItems: "center"}}>
-            <MainCard />
-            <h2>An example of plant care instructions:</h2>
-            <Plant />
-          </Box>
-        </main>
+            <Header />
 
-      <RegisterForm onSubmit={handleRegisterSubmit} />
-        <LoginForm onSubmit={handleLoginSubmit} />
-    </>
-  );
+            <div>
+                <Switch>
+                    <Route exact path="/">
+                        <MainCard />
+                    </Route>
+
+                    <Route path="/register">
+                        <RegisterForm onSubmit={handleRegisterSubmit} />
+                    </Route>
+
+                    <Route path="/login">
+                        <LoginForm onSubmit={handleLoginSubmit} />
+                    </Route>
+                </Switch>
+            </div>
+        </>
+    );
 }
 
 export default App;
