+++ plants/src/assets/mainPlant.jpg
