+++ plants/src/components/MainCard.js
