+++ plants/src/components/Plant.js
